Example data

Example data used to display the features of the traj package.

example.data

Format

List of 2 data.frames:

$ data: 130 obs. of 7 variables.X1 to X6 correspond to the 6 measurements obtained on the 130 subjects: html
ID: num 1 2 3 4 5 6 7 8 9 10 ... X1: num 5.66 23.59 15.47 7.31 12.84 ... X2: num 9.34 11.75 8.76 11.69 11.09 ... X3: num 3.77 7.68 6.49 12.48 7.65 ... X4: num 17.36 12.83 11.26 8.89 10.27 ... X5: num 8.82 13 10.42 6.52 12.45 ... X6: num 9.28 9.66 17.41 7.7 11.56 ... html
$ time: 130 subjects with 7 variables. time1 to time6 correspond to the measurement times for the variables X1 to X6: html
ID : num 1 2 3 4 5 6 7 8 9 10 ... time.1: num 1 1 1 1 1 1 1 1 1 1 ... time.2: num 2 2 2 2 2 2 2 2 2 2 ... time.3: num 3 3 3 3 3 3 3 3 3 3 ... time.4: num 4 4 4 4 4 4 4 4 4 4 ... time.5: num 5 5 5 5 5 5 5 5 5 5 ... time.6: num 6 6 6 6 6 6 6 6 6 6 ... html

Author

Marie-Pierre Sylvestre, Dan Vatnik

marie-pierre.sylvestre@umontreal.ca

Examples

if (FALSE) {
# data and time
data = example.data$data 
time = example.data$time
}