Plot Median Trajectory of traj Object

Plot cluster-specific median trajectory for one or all clusters provided by a traj object.

plotMedTraj(
  x,
  clust.num = NULL,
  plot.percentile = TRUE,
  low.percentile = 0.1,
  high.percentile = 0.9,
  ...
)

Arguments

x

traj object.

clust.num

Integer indicating the cluster number to plot.NULL to print all clusters. Defaults to NULL.

plot.percentile

Value indicating if the function should plot percentiles. Defaults to TRUE.

low.percentile

Value of the lower percentile to be plotted. Must be between 0 and 1. Defaults to 0.1.

high.percentile

Value of the high percentile to be plotted. Must be between 0 and 1. Defaults to 0.9.

...

Extra parameters used in the plot function.

Details

The function plots cluster specific median trajectory calculated at each time point, in addition to 10th and 90th percentiles. By setting the clust.num argument to an integer corresponding to a cluster number, one can plot the median trajectory of that cluster only. Any other plotting arguments can be added to the function.

See also

Author

Marie-Pierre Sylvestre, Dan Vatnik

marie-pierre.sylvestre@umontreal.ca

Examples

if (FALSE) {
# Setup data and time
data = example.data$data
time = example.data$time

# Run step1measures, step2factors and step3clusters with a predetermined number of clusters
s1 = step1measures(data,time, ID=TRUE)
s2 = step2factors(s1)
s3.4clusters = step3clusters(s2, nclust = 4)

# Plot median trajectories
plotMedTraj(s3.4clusters)
}