Wrapper Function to Perform Trajectory Analysis

Run three steps of trajectory analysis with default parameters.

wrapperTraj(Data, ID = FALSE)

Arguments

Data

Data frame containing trajectory data. Each line should contain sequential observations. See details and help file for the step1measures function.

ID

Logical. Set to FALSE if the first column of Data corresponds to an ID variable. Defaults to FALSE.

Value

The result is a traj object. Details can be found in the 'Value' section of step3clusters.

Details

The function runs the full three step trajectory analysis and returns a traj object. It will execute step1measures, step2factors and step3clusters sequentially with their default parameters. The result of step3clusters will be returned. Details regarding the data and the time arguments are found in the 'Details' section of step1measures.

Author

Marie-Pierre Sylvestre, Dan Vatnik

marie-pierre.sylvestre@umontreal.ca

Examples

if (FALSE) {
# Setup data
data = example.data$data

# Run clustering wrapper function
wt = wrapperTraj(data, ID = TRUE)

# Display and plot "traj" object
wt

summary(wt)

plot(wt)
}